backend/requirements.txt
flask>=3.0.3
flask-cors>=4.0.1
flask-socketio>=5.3.6
python-socketio>=5.11.3
kafka-python>=2.0.2
gunicorn>=22.0.0
PyJWT>=2.8.0
